requirements.txt
APScheduler==3.10.4
backports.zoneinfo==0.2.1;python_version<"3.9"
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
flask==3.0.3
Flask-Cors==4.0.0
# Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
# mysql-connector-python==9.0.0
# mysqlclient==2.2.4
numpy==1.24.4
pandas==2.0.3
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.29
typing-extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
werkzeug==3.0.2
zipp==3.18.1
# pymysql
psycopg2